+++ src/compositor.rs
@@ -289,6 +289,7 @@ fn start_compositor2(
         color_manager,
         float_above_fullscreen: Cell::new(false),
         icons: Default::default(),
+        show_pin_icon: Cell::new(false),
     });
     state.tracker.register(ClientId::from_raw(0));
     create_dummy_output(&state);
@@ -618,6 +619,7 @@ fn create_dummy_output(state: &Rc<State>) {
         tray_start_rel: Default::default(),
         tray_items: Default::default(),
         ext_workspace_groups: Default::default(),
+        pinned: Default::default(),
     });
     let dummy_workspace = Rc::new(WorkspaceNode {
         id: state.node_ids.next(),